requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
astroid==2.15.6
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
dill==0.3.7
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
platformdirs==3.10.0
pylint==2.17.5
python-dotenv==1.0.0
python-telegram-bot==20.5
requests==2.31.0
sniffio==1.3.0
texttable==1.6.7
tomlkit==0.12.1
urllib3==2.0.4
wrapt==1.15.0
yarl==1.9.2
